fix casting warning in pagetable c

+++ pagetable.c
@@ -162,7 +162,7 @@ char *find_physpage(addr_t vaddr, char type) {
     if(!(pgdir[idx].pde & PG_VALID)){
         pgdir[idx] = init_second_level();
     }
-    
+
     pgtbl_entry_t *pgtbl;
     pgtbl = (pgtbl_entry_t*) (pgdir[idx].pde & PAGE_MASK);
 	// Use vaddr to get index into 2nd-level page table and initialize 'p'
@@ -171,12 +171,12 @@ char *find_physpage(addr_t vaddr, char type) {
 
 	// Check if p is valid or not, on swap or not, and handle appropriately
 	int frame_no;
-    
+
     // if VALID is 1, page is in memory. it's a hit
-    if ((p->frame) & PG_VALID) { 
+    if ((p->frame) & PG_VALID) {
 		hit_count ++;
-        frame_no = p->frame >> PAGE_SHIFT; 	
-        p->frame |= CLOCK_REF; // set CLOCK_REF bit to 1 for clock algo	
+        frame_no = p->frame >> PAGE_SHIFT;
+        p->frame |= CLOCK_REF; // set CLOCK_REF bit to 1 for clock algo
 	}
 	// else, page is not in memory. it's a miss
 	else {
@@ -186,10 +186,10 @@ char *find_physpage(addr_t vaddr, char type) {
 		if(frame_no == -1){
             perror("error allocating_frame");
         }
-        
+
         p->frame &= PAGE_SHIFT-1;
         p->frame |= frame_no << PAGE_SHIFT;
-        
+
 		//ONSWAP is 1, page is on disk
 		if ((p->frame) & PG_ONSWAP) {
 			swap_pagein(frame_no, p->swap_off);
@@ -197,7 +197,7 @@ char *find_physpage(addr_t vaddr, char type) {
 			p->frame |= PG_VALID;
 		}
 		// page is not on disk, first time access the page
-		else {			
+		else {
 			init_frame(frame_no, vaddr);
 			//mark it as DIRTY, so it will be written to swap when evicted
 			p->frame |= PG_DIRTY;
@@ -217,10 +217,10 @@ char *find_physpage(addr_t vaddr, char type) {
     char *mem_ptr = &physmem[frame_no*SIMPAGESIZE];
     addr_t *vaddr_ptr = (addr_t *)(mem_ptr + sizeof(int));
     *vaddr_ptr = vaddr;
-    
+
 	// Call replacement algorithm's ref_fcn for this page
 	ref_fcn(p);
-	
+
 
 
 	// Return pointer into (simulated) physical memory at start of frame
@@ -254,7 +254,7 @@ void print_pagetbl(pgtbl_entry_t *pgtbl) {
 				printf("in frame %d\n",pgtbl[i].frame >> PAGE_SHIFT);
 			} else {
 				assert(pgtbl[i].frame & PG_ONSWAP);
-				printf("ONSWAP, at offset %lld\n",pgtbl[i].swap_off);
+				printf("ONSWAP, at offset %i\n",(int)pgtbl[i].swap_off);
 			}
 		}
 	}
